requirements.txt
Flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
transformers==4.35.2
torch
pdfplumber
python-docx
PyPDF2
gunicorn
